+++ CHANGELOG.md
@@ -79,6 +79,7 @@
   - `cowel_var_get`
   - `cowel_var_let`
   - `cowel_var_set`
+- add a `cowel_as_text` content policy directive (#202)
 - add the ability to store groups (and generally not just strings) in variables (#73, #166)
 - add support for `abbreviation` and `figment` aliases to `cowel_char_by_name` (#11)
 - add language extension for VSCode (#140)